% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_reader.R
\name{directory_reader_csv2}
\alias{directory_reader_csv2}
\title{Directory reader csv2 files}
\usage{
directory_reader_csv2(folder)
}
\arguments{
\item{folder}{The path to the folder containing csv files.}
}
\value{
a single dataframe
}
\description{
This function returns a single dataframe based on the identical csv files in the designated folder.
}
