% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{directoryreader-package}
\alias{directoryreader-package}
\alias{directoryreader}
\title{directory reader}
\description{
This package contains functions to return a single dataframe based on the identical files in the designated folder.
}
\author{
Adriaan Nering Bogel
}
